backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
openpyxl==3.1.5
reportlab==4.2.5
python-dateutil==2.9.0
